Updates to ldap module to improve user syncing in large environments (#18)
Updates to `ldap` module to improve user syncing in large environments (#18)


# API Changes:
- Update `LDAPService` to always use the new `LDAPIterator` (see below) for calls to `LDAPService->getUsers()`. This ensures that all users will be returned, rather than only the first list of users (up to the LDAP maximum page size as was previously occurring).

- Deprecate memberLoggedIn() since it is now unused. Use afterMemberLoggedIn instead.


# Notable features:
## New way of syncing user and group data
Adds a new `LDAPIterator` class that provides the ability to return all objects that match a search filter, regardless of the LDAP servers max page size.

Note: There shouldn't be any breaking changes as the only option that we don't pass in is a search scope, and this is fine because the hard-coded scope (`Ldap::SEARCH_SCOPE_SUB`) ended up calling `ldap_search()` which the new iterator still does.

Note 2: Many thanks to GitHub user ThaDafinser for providing the initial implementation: https://gist.github.com/ThaDafinser/1d081bed8e5e6505e97bedf5863a187c

## Allow SilverStripe attributes to be reset by AD

Attributes missing from AD are not reset by the module, which can be a business case. This feature adds the ability for attributes to be reset if they are not present in the LDAP source data but present in the mapping.

This feature is opt-in as it's destructive (attributes missing from LDAP for a given user will be removed from the SilverStripe database).

## New method for photo thumbnail generation
Photos were not being saved previously, as the original module code was not updated to account for SilverStripe 4's new asset abstraction work. This has now been updated, along with some tweaks to ensure that the filesystem is not thrashed heavily.

LDAP sync works by syncing all users on a regular schedule (e.g. daily or hourly). Because LDAP does not natively support requesting 'only objects that have changed', we need to sync all users every time. This leads to thumbnail photos being marked as having changed every time the sync task runs for all users.

Instead of this, we now prevent this file version thrashing by comparing the binary blobs and only updating the file record if the files differ.

- Resolve issues with thumbnail storage (photos were not always being stored, with 'Extension is not allowed' errors in rare occurrences (~1 in 1,000), even though the extension is always set to .jpg)
- Prevent File ID thrashing, by updating the existing image object with the new image binary blob, instead of deleting the old File object and creating a new one (also by overwriting the old image instead of creating a new file, prevents `vN` from being appended to the end of the filename every time)
- Dont overwrite images that already exist already: When syncing large data sets, there's no point in overwriting images if it is the same image. Instead, check the file hash of the old and new, and overwrite if there's changes. For large LDAP sources, this resolves an issue where the syncs are incredibly file intensive.

## Resolve removed LDAP group mappings

Members which belonged to LDAP mapped groups weren't being removed from these associations if the LDAP data no longer contained the mappings. This meant if a group association was removed from LDAP the member was never updated.

This also accounts for situations where there's no more mappings or no mappings are given.

## Add extension hooks
Extension hooks are added for `onAfterLDAPMemberSyncTask` and `onAfterLDAPGroupSyncTask` to allow app-specific code to run after all members and groups are synced (for example, to map users by FK IDs instead of DN attribute values)


# Minor features:
- Add ability to get to the underlying `Zend\Ldap\Ldap` object to gain access to the PHP `ldap` resource

- Add new `BuildTask` that syncs just one LDAP user based on email address. This can be used to test LDAP syncing (make sure you run the `LDAPMigrateExistingMembersTask` task first if the user already exists in SS to map the GUID across)

- Properly save thumbnail photos (old code hadn't changed from the SS3 version and no longer worked with the new SS4 asset abstraction)

- Allow automatic update from LDAP when a member logs in to fail without breaking login based on configuration

- Bump minimum core version to 4.1


# Bug Fixes:
- Fix to use `afterMemberLoggedIn`, as the `memberLoggedIn` method is no longer called in core.
- Fix `LDAPDebugController::Users()` to not require so much memory by retrieving all attributes for all users (an issue now that we could return significantly more than 1,000 users)
- Set memory limit to unlimited in `LDAPDebugController`, allowing for massive resultsets
- Reset the pagination control after performing a paginated search, to allow additional searches to work
- Fix `LDAPAllSyncJob` to sync both groups and members, not just groups twice
- Set the max execution time to 60 minutes for user syncing and 15 minutes for group syncing, which can take a while under normal circumstances with large amounts of groups or users to sync
- Up memory limit to 1GB to account for massive datasets (LDAP doesn't support generators and pagination fails quickly as on busy LDAP servers pagination control cookies can get dropped from cache quickly, so we have to load all LDAP users into memory before processing any)
- Add `cn` in to the list of standard attributes to pull so we can use this in the log message
- Fix `LDAPMemberSyncOneTask` to fail quickly if no user is found in LDAP
- Fix minor security risk by not using the SAM Account Name (aka username) in the filename, instead using the user's GUID
- Update LDAPIterator::__construct docblock and constructor to support PHP5.6
- Fix issue with postgres compatibility: table identifiers not quoted
- Remove PHP 5 incompatible scalar type hinting
- Update config->get() style and remove unnecessary else clause, optimise imports
- Promote processSearchResults to protected
- Prefer getGateway() over direct property access, promote processThumbnailPhoto() visibility
- Use Injector to load a new Image instance
- Add public setter for LDAPService
- Use LDAPService setter, change die() for exit() and import missing Exception namespace
- Fix issue with postgres compatibility, unquoted column name
- Remove redundant limit(1) on DataList where first() is called afterwards
- Implement public setter for LDAPService in LDAPMemberSyncTask


# Testing: 
- Run automated phpcbf linting with slightly modified PSR-2 standard
- Fix code linting errors preventing successful build
- Add various core recipe versions to the Travis builds, add PHP 7.2 and 7.3
- Fix signature compatibility with DataObject::write() in SS 4.3 and remove ldap from Travis (installed by default)


# Documentation:
- Add documentation for using the new `LDAPIterator`
- Update developer documentation with corrected value for `groups_search_locations` YML fragment
- Add documentation for new features added to the module (`onAfterLDAPMemberSyncTask` and `onAfterLDAPGroupSyncTask`)
- Fix spelling mistakes in documentation
- Update comments to be technically correct (compared to the code they describe)